Added the option to return a data.frame of ranges
If arg return.range is TRUE, the function searches for and returns the minimum and maximum latitude and longitude.  These are returned in a data.frame with a row for each requested record.
A problem with 0 Lat and 0/180 Long may occur since these values are not necessarily followed by N/S or E/W, respectively.  Need to re-write to deal with these cases.